mm: multi-gen LRU: don't sync disk for each aging cycle
wakeup_flusher_threads() was added under the assumption that if a
system runs out of clean cold pages, it might want to write back dirty
pages more aggressively so that they can become clean and be dropped.

However, doing so can breach the rate limit a system wants to impose
on writeback, resulting in early SSD wearout.

Reported-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: Yu Zhao <yuzhao@google.com>